Fix signed integer overflow in quantileExactExclusive/Inclusive interpolation
QuantileExactExclusive::getFloat, QuantileExactExclusive::getManyFloat and
QuantileExactInclusive::getFloat computed the interpolation delta as
`*nth_elem - array[n - 1]` in the value's own integer type and only then cast
the result to Float64. For Int64 inputs spanning a large range (e.g. a value
near INT64_MAX and a negative value) this subtraction overflows signed int64,
which is undefined behavior under UBSan and, in release builds, wraps to a
garbage value producing a wrong quantile outside the data range.

Cast each operand to Float64 before subtracting, so the delta is computed in
floating point and cannot overflow. This matches the already-correct sibling
QuantileExactInclusive::getManyFloat.

Found by the AST fuzzer (arm_asan_ubsan):
src/AggregateFunctions/QuantileExact.h:155:120: runtime error: signed integer
overflow: 9223372036854775807 - -2147483647 cannot be represented in type 'long'

+++ src/AggregateFunctions/QuantileExact.h
@@ -152,7 +152,7 @@ struct QuantileExactExclusive : public QuantileExact<Value>
             ::nth_element(array.begin(), array.begin() + n - 1, array.end());
             auto * nth_elem = std::min_element(array.begin() + n, array.end());
 
-            return static_cast<Float64>(array[n - 1]) + (h - static_cast<Float64>(n)) * static_cast<Float64>(*nth_elem - array[n - 1]);
+            return static_cast<Float64>(array[n - 1]) + (h - static_cast<Float64>(n)) * (static_cast<Float64>(*nth_elem) - static_cast<Float64>(array[n - 1]));
         }
 
         return std::numeric_limits<Float64>::quiet_NaN();
@@ -181,7 +181,7 @@ struct QuantileExactExclusive : public QuantileExact<Value>
                     ::nth_element(array.begin() + prev_n, array.begin() + n - 1, array.end());
                     auto * nth_elem = std::min_element(array.begin() + n, array.end());
 
-                    result[indices[i]] = static_cast<Float64>(array[n - 1]) + (h - static_cast<Float64>(n)) * static_cast<Float64>(*nth_elem - array[n - 1]);
+                    result[indices[i]] = static_cast<Float64>(array[n - 1]) + (h - static_cast<Float64>(n)) * (static_cast<Float64>(*nth_elem) - static_cast<Float64>(array[n - 1]));
                     prev_n = n - 1;
                 }
             }
@@ -216,7 +216,7 @@ struct QuantileExactInclusive : public QuantileExact<Value>
             ::nth_element(array.begin(), array.begin() + n - 1, array.end());
             auto * nth_elem = std::min_element(array.begin() + n, array.end());
 
-            return static_cast<Float64>(array[n - 1]) + (h - static_cast<Float64>(n)) * static_cast<Float64>(*nth_elem - array[n - 1]);
+            return static_cast<Float64>(array[n - 1]) + (h - static_cast<Float64>(n)) * (static_cast<Float64>(*nth_elem) - static_cast<Float64>(array[n - 1]));
         }
 
         return std::numeric_limits<Float64>::quiet_NaN();
